schnauzer: refactor stdlib helpers to new module
